Yakuza 6 autostart, autosplit, all game versions (#2)
* Autosplitter overhaul (GOG version only)

* Added versions: M Store, Steam launch, Steam current patch

* Full support for all Steam versions

Since every version after launch occupies the same size in memory, instead we'll compare the TimeDateStamp in the COFF header of the EXE, which is the file creation date.

* Convert tabs to spaces to fix weird formatting issue

* Updated vars.OngoingFight logic

Checking if current.Magic has changed doesn't work soon enough with this particular pointer (it worked with a prior one I had tested).

Instead, we'll check the absolute difference in IGT from frame to frame during a fight. IGT is set to the new value (either on continue or loading a save) quickly enough to be more valuable.

* A few more fight splits

* More misc. splits, and autosplitting functionality for them

* Swapped Poem Stones for Yamano Dodai's Grave

* Autoreset double-check

"boot" always hits in my experience, but it's fine to be thorough.

* Millennium Tower fix + Ch.3 Baby split change

The Millennium Tower setpiece string had a typo on my part.

The Ch.3 baby segment now more naturally splits after returning to the apartment for the night, rather than when taking the taxi to the baseball diamond.

* Autostart for NG+ runs

I started by adding a setting for NG+ that, when enabled, would check for starts less leniently. But that allows some file loads from the main menu to false start, and although those false starts could be immediately caught and auto-reset, that's still not ideal.

This solution should work perfectly, in combination with the refresh rate being set to 30. With refresh at 60, we would catch a false stop in IGT every second, probably because the game was programmed for 30.

* Fix: False splits in boss fights

The enemy count (very rarely) flickers to 0 for less than a single frame. This may be a quirk of the game's conversion from 30 FPS, but at any rate, it's impossible to predict when it will occur.

This fix checks two separate enemy count values in memory for agreement before splitting. In hours of testing, I haven't seen two values flicker simultaneously (I checked 22 pointers concurrently, actually), as it's a rare phenomenon to begin with. There are two or three other regions in memory that can be piled onto this check, but that's almost certainly unnecessary.

* Removed auto-reset

No one will use it anyway!